requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
base58==2.1.1
beautifulsoup4==4.12.2
billiard==4.1.0
bitarray==2.8.1
build==1.0.3
cachetools==5.3.2
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==41.0.4
cytoolz==0.12.2
daphne==4.0.0
distlib==0.3.7
Django==4.2.3
django-amis-render==0.0.8
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-js-asset==2.1.0
django-mptt==0.15.0
django-multiselectfield==0.1.12
django-simpleui==2023.3.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
eth-account==0.11.0
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==4.0.0
eth_abi==5.0.1
exceptiongroup==1.1.3
filelock==3.12.2
flower==2.0.1
frozenlist==1.4.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-texttospeech==2.14.2
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
h2==4.1.0
hexbytes==0.3.1
hpack==4.0.0
humanize==4.8.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==7.0.1
incremental==22.10.0
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
lru-dict==1.2.0
lxml==4.9.3
MarkupSafe==2.1.3
msgpack==1.0.6
multiaddr==0.0.9
multidict==6.0.4
mysqlclient==2.2.0
netaddr==0.9.0
numpy==1.25.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parsimonious==0.9.0
pip-tools==7.3.0
platformdirs==3.10.0
priority==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pylru==1.2.1
pyOpenSSL==23.2.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pysha3==1.0.2
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-memcached==1.59
pytz==2023.3
pyunormalize==15.0.0
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rlp==4.0.0
rpds-py==0.9.2
rsa==4.9
selenium==4.11.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
trio==0.22.2
trio-websocket==0.10.3
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
uWSGI==2.0.22
varint==1.0.2
vine==5.0.0
virtualenv==20.24.3
wcwidth==0.2.6
web3==6.15.1
webdriver-manager==4.0.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
zope.interface==6.0
